partial-fix for mapping of solutions
FIXME we currently only support either all exercises having a solution
or none of them having one, as the post-hook needs to be executed
exactly once. This can either be after the solution is read, or after
the exercise is read. To fix the, we could either check whether the post
hook of a previous exercise was run when starting a new sibling (i.e. by
always setting a dirty flag in the post-hook, but only running the
clean-up code in the pre-hook of a sibling if we are dirty) or using
different global variable for nesting vs. solutions (or maybe even local
variables).